[NEW] Menu and Optimizer
The menu and the Optimizer got some new styles that will help with the implementation of some new features and you can see some of them right away.

Added a "Toolbox" button to the menu that will help with managing some challenges, like deleting all of them, deleting all the completed and so on.