[REWORK] Optimizer
Due to feedbacks, the methodology behind the Optimizer is now changed.

It will basically, show to you EVERY CHALLENGE that has 2 occurrences or more and not just show you the challenge that appear the most.

It's on accordion format and all styles are still being picked, most of the stuff you will see here it's placeholder.

There will be bugs.
Or no.

(probably yes)