Changes on the API usage and Styling
The mobile menu was changed due to refactoring of the code.

Note about the API change:

A "manual" timer based on the initial map duration was added to avoid every second of a map rotation being a fetch. This will make the timers more accurate and less intensive server-side. The only problem about this approach, is that, the code behind this method, stops when the browser is idle.
So, let's say that the user is on ALT TAB for a while, when they go back, it may be on a wrong timer and need to refresh the page to get the correct one. I'm thinking on a solution to fix this problem, but I don't think many users will afk on the rotations page anyway, so it's pretty much usable.